Add input field for maximum number of transcripts, change styling of the input page, added more checks for uploaded file and more precise error messages.
Maximum number of inputs can now be selected via a slider or the input field. The files is now checked in more ways and processed a bit more efficiently. This should catch all the wrong possible inputs and show an corresponding error message without throwing a server error. Changed the styling of the input page a bit and added "alerts"